student_management_sys/requirements.txt
aioredis==1.3.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==3.0.5
channels-redis==3.4.0
constantly==23.10.4
cryptography==42.0.2
daphne==3.0.2
decorator==5.1.1
Django==5.0.1
django-cleanup==8.0.0
django-extensions==3.2.3
djangochannelsrestframework==1.0.0
djangorestframework==3.14.0
executing==2.0.1
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
ipython==8.20.0
jedi==0.19.1
matplotlib-inline==0.1.6
msgpack==1.0.7
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==24.0.0
pypng==0.20220715.0
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
Unidecode==1.3.8
wcwidth==0.2.13
zope.interface==6.1
